requirements.txt
streamlit>=1.24.0
requests>=2.31.0
pandas>=2.0.3
Pillow>=10.0.0
streamlit-folium>=0.11.0
folium>=0.14.0
sentence-transformers>=2.2.2
transformers>=4.33.3
torch>=2.0.1
numpy>=1.24.4
python-dotenv>=1.0.0